# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EngHub - Engineering Projects Showcase API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface UserRegistration {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  courseName?: string | null;
  /** @nullable */
  teamMembers?: string | null;
  /** @nullable */
  componentsTags?: string | null;
  /** @nullable */
  costLevel?: string | null;
  /** @nullable */
  difficultyLevel?: string | null;
  /** @nullable */
  inspiredByLink?: string | null;
  /** @nullable */
  challengesText?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  reportLink?: string | null;
  /** @nullable */
  videoLink?: string | null;
  upvotes: number;
  userId: number;
  /** @nullable */
  authorUsername?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  courseName?: string;
  teamMembers?: string;
  componentsTags?: string;
  costLevel?: string;
  difficultyLevel?: string;
  inspiredByLink?: string;
  challengesText?: string;
  imageUrl?: string;
  reportLink?: string;
  videoLink?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  courseName?: string;
  teamMembers?: string;
  componentsTags?: string;
  costLevel?: string;
  difficultyLevel?: string;
  inspiredByLink?: string;
  challengesText?: string;
  imageUrl?: string;
  reportLink?: string;
  videoLink?: string;
}

export interface Comment {
  id: number;
  text: string;
  userId: number;
  projectId: number;
  /** @nullable */
  authorUsername?: string | null;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  text: string;
}

export interface PlatformStats {
  totalProjects: number;
  totalUsers: number;
  totalComments: number;
  totalUpvotes: number;
}

export interface CourseCount {
  courseName: string;
  count: number;
}

export type ListProjectsParams = {
/**
 * Filter by course name
 */
course_name?: string;
/**
 * Filter by difficulty level
 */
difficulty_level?: string;
/**
 * Filter by cost level
 */
cost_level?: string;
/**
 * Search by title or description
 */
search?: string;
/**
 * Sort order
 */
sort?: ListProjectsSort;
};

export type ListProjectsSort = typeof ListProjectsSort[keyof typeof ListProjectsSort];


export const ListProjectsSort = {
  newest: 'newest',
  upvotes: 'upvotes',
} as const;