# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EngHub - Engineering Projects Showcase API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: projects
    description: Project management
  - name: comments
    description: Comment management
  - name: stats
    description: Platform statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: course_name
          schema:
            type: string
          description: Filter by course name
        - in: query
          name: difficulty_level
          schema:
            type: string
          description: Filter by difficulty level
        - in: query
          name: cost_level
          schema:
            type: string
          description: Filter by cost level
        - in: query
          name: search
          schema:
            type: string
          description: Search by title or description
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, upvotes]
          description: Sort order
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{id}/upvote:
    post:
      operationId: upvoteProject
      tags: [projects]
      summary: Upvote a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Upvoted project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/platform:
    get:
      operationId: getPlatformStats
      tags: [stats]
      summary: Get platform-wide statistics
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

  /stats/trending:
    get:
      operationId: getTrendingProjects
      tags: [stats]
      summary: Get top trending projects by upvotes
      responses:
        "200":
          description: Trending projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /stats/courses:
    get:
      operationId: getCourseBreakdown
      tags: [stats]
      summary: Get project count breakdown by course
      responses:
        "200":
          description: Course breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    UserProfile:
      type: object
      required: [id, username, email, createdAt]
      properties:
        id:
          type: number
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string

    UserRegistration:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6

    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    Project:
      type: object
      required: [id, title, description, upvotes, userId, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        courseName:
          type: ["string", "null"]
        teamMembers:
          type: ["string", "null"]
        componentsTags:
          type: ["string", "null"]
        costLevel:
          type: ["string", "null"]
        difficultyLevel:
          type: ["string", "null"]
        inspiredByLink:
          type: ["string", "null"]
        challengesText:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        reportLink:
          type: ["string", "null"]
        videoLink:
          type: ["string", "null"]
        upvotes:
          type: number
        userId:
          type: number
        authorUsername:
          type: ["string", "null"]
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        courseName:
          type: string
        teamMembers:
          type: string
        componentsTags:
          type: string
        costLevel:
          type: string
        difficultyLevel:
          type: string
        inspiredByLink:
          type: string
        challengesText:
          type: string
        imageUrl:
          type: string
        reportLink:
          type: string
        videoLink:
          type: string

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        courseName:
          type: string
        teamMembers:
          type: string
        componentsTags:
          type: string
        costLevel:
          type: string
        difficultyLevel:
          type: string
        inspiredByLink:
          type: string
        challengesText:
          type: string
        imageUrl:
          type: string
        reportLink:
          type: string
        videoLink:
          type: string

    Comment:
      type: object
      required: [id, text, userId, projectId, createdAt]
      properties:
        id:
          type: number
        text:
          type: string
        userId:
          type: number
        projectId:
          type: number
        authorUsername:
          type: ["string", "null"]
        createdAt:
          type: string

    CommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1

    PlatformStats:
      type: object
      required: [totalProjects, totalUsers, totalComments, totalUpvotes]
      properties:
        totalProjects:
          type: number
        totalUsers:
          type: number
        totalComments:
          type: number
        totalUpvotes:
          type: number

    CourseCount:
      type: object
      required: [courseName, count]
      properties:
        courseName:
          type: string
        count:
          type: number