# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EngHub - Engineering Projects Showcase API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  text: string;
  userId: number;
  projectId: number;
  /** @nullable */
  authorUsername?: string | null;
  createdAt: string;
}