# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EngHub - Engineering Projects Showcase API
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /** @minLength 1 */
  text: string;
}