# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EngHub - Engineering Projects Showcase API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalProjects: number;
  totalUsers: number;
  totalComments: number;
  totalUpvotes: number;
}