# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EngHub - Engineering Projects Showcase API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  courseName?: string;
  teamMembers?: string;
  componentsTags?: string;
  costLevel?: string;
  difficultyLevel?: string;
  inspiredByLink?: string;
  challengesText?: string;
  imageUrl?: string;
  reportLink?: string;
  videoLink?: string;
}